# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a backend server for sylu-oj.",
        "title": "sylu-oj",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/exam/add_exam": {
            "post": {
                "description": "创建实验",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "exam message",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.AddExam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/error.ResponseNormal"
                        }
                    }
                }
            }
        },
        "/question/add_question": {
            "post": {
                "description": "发布题目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "question message",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.AddQuestion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/error.ResponseNormal"
                        }
                    }
                }
            }
        },
        "/question/commit_answer": {
            "post": {
                "description": "查看题目详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "commit your answer of question",
                        "name": "answer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Answer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/error.ResponseNormal"
                        }
                    }
                }
            }
        },
        "/question/del": {
            "delete": {
                "description": "删除题目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "question id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/error.ResponseNormal"
                        }
                    }
                }
            }
        },
        "/question/get_question_msg": {
            "get": {
                "description": "查看题目详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "get question message",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/error.ResponseNormal"
                        }
                    }
                }
            }
        },
        "/question/list": {
            "get": {
                "description": "查看题目列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "number",
                        "description": "page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "rows per page",
                        "name": "number",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "question title",
                        "name": "search_title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "question tag",
                        "name": "search_tag",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "question degree",
                        "name": "search_degree",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "reorder",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/error.ResponseNormal"
                        }
                    }
                }
            }
        },
        "/test/ping": {
            "get": {
                "description": "测试连接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "测试接口",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/error.ResponseNormal"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "用户登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "user message",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/error.ResponseNormal"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "get": {
                "description": "退出登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/error.ResponseNormal"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "user message",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.UserRegister"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/error.ResponseNormal"
                        }
                    }
                }
            }
        },
        "/utils/email_verify": {
            "get": {
                "description": "发送邮箱验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "邮箱验证接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/error.ResponseNormal"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "error.ResponseNormal": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回内容"
                },
                "msg": {
                    "description": "返回消息",
                    "type": "string"
                }
            }
        },
        "schema.AddExam": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "student": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "schema.AddQuestion": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "degree": {
                    "type": "integer"
                },
                "expected_output": {
                    "type": "string"
                },
                "input_test": {
                    "type": "string"
                },
                "tag": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "schema.Answer": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "schema.UserLogin": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "schema.UserRegister": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "verify_code": {
                    "type": "string"
                }
            }
        }
    }
}